# swagger.yaml
openapi: "3.0.0"
info:
  title: "My API"
  description: "API documentation with Swagger"
  version: "1.0.0"
servers:
  - url: "http://localhost:3000"

paths:
  /api/items:
    post:
      summary: "Create a new item"
      requestBody:
        description: "Item object that needs to be added"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        201:
          description: "Item created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string

    get:
      summary: "Get all items"
      responses:
        200:
          description: "A list of items"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string

  /api/items/{id}:
    get:
      summary: "Get an item by ID"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "Item found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        404:
          description: "Item not found"

    put:
      summary: "Update an item by ID"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        description: "Updated item object"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        200:
          description: "Item updated successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        404:
          description: "Item not found"

    delete:
      summary: "Delete an item by ID"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        204:
          description: "Item deleted successfully"
        404:
          description: "Item not found"